elephant_movement_ai/requirements.txt
# Mathematics & Data Processing
numpy
scipy
pandas

# Machine Learning & RL
scikit-learn
torch
stable-baselines3
gymnasium

# GIS & Spatial Analytics
geopandas
shapely

# Visualization
matplotlib
plotly
folium

# Type hints and standard tooling
typing-extensions
pyyaml
